refactor: Layered Architecture 적용 및 코드 품질 개선
## 📦 패키지 구조 리팩토링 (DRY 원칙 적용)

**Before (DRY 원칙 위반):**
```
infrastructure/client/
├── binance/
│   ├── config/      ← 중복
│   ├── exception/   ← 중복
│   └── interceptor/ ← 중복
└── coinone/
    ├── config/      ← 중복
    ├── exception/   ← 중복
    └── interceptor/ ← 중복
```

**After (DRY 원칙 준수):**
```
infrastructure/
├── config/
│   ├── binance/     # Binance API 설정
│   └── coinone/     # Coinone API 설정
├── exception/
│   ├── binance/     # Binance 예외 처리
│   └── coinone/     # Coinone 예외 처리
└── interceptor/
    ├── common/      # 공통 인터셉터
    ├── binance/     # Binance 인증
    └── coinone/     # Coinone 인증
```

## 🔧 주요 변경사항

### 1. Properties 클래스 개선
- BinanceProperties, CoinoneProperties에 @Validated 어노테이션 추가
- 필수 필드에 @NotBlank/@NotNull validation 추가
- 애플리케이션 시작 시 설정값 검증으로 런타임 에러 방지

### 2. 버그 수정
- CoinoneOrderCancleRequest → CoinoneOrderCancelRequest 클래스명 오타 수정
- 모든 import 구문을 새로운 패키지 구조에 맞게 업데이트

### 3. 문서화 개선
- README.md 대폭 확장 (프로젝트 구조, 사용법, 보안 가이드 등)
- 모든 클래스에 포괄적인 JavaDoc 추가
- 설정 예시 및 API 문서 링크 제공

### 4. 테스트 코드 정리
- 테스트 클래스들의 패키지 경로를 새로운 구조에 맞게 이동
- MockCoinoneTickerApiClient 등 테스트 지원 클래스 경로 수정

## 📈 기대 효과

- **확장성**: 새로운 거래소 추가 시 기존 구조 활용 가능
- **유지보수성**: 공통 기능과 거래소별 특화 기능 명확한 분리
- **안정성**: validation으로 설정 오류 사전 방지
- **개발자 경험**: 상세한 문서화로 온보딩 시간 단축
