+++ src/main/java/sungjun/bitcoin/algorithmtrading/infrastructure/client/binance/BinanceAccountApiClient.java
@@ -21,59 +21,49 @@ import sungjun.bitcoin.algorithmtrading.infrastructure.client.binance.response.B
 @HttpExchange
 public interface BinanceAccountApiClient {
 
-    /**
-     * 계정 정보를 조회합니다 (전체 파라미터 버전).
-     * <p>
-     * 사용자의 계정 잔고, 거래 수수료 정보 등을 포함한 상세 계정 정보를 조회합니다.
-     * </p>
-     *
-     * @param omitZeroBalances 잔고가 0인 자산을 응답에서 제외할지 여부
-     * @param recvWindow 요청 유효 시간 (밀리초, 최대 60000)
-     * @param timestamp 요청 시각 (밀리초 단위 Unix timestamp)
-     * @return {@link BinanceAccountApiResponse} 계정 정보 응답
-     * @throws sungjun.bitcoin.algorithmtrading.infrastructure.client.binance.exception.BinanceApiException API 호출 실패 시
-     */
+    /****
+                       * Retrieves detailed account information from Binance, including balances and trading fee details.
+                       *
+                       * @param omitZeroBalances whether to exclude assets with zero balance from the response
+                       * @param recvWindow the validity window for the request in milliseconds (maximum 60000)
+                       * @param timestamp the Unix timestamp in milliseconds representing the request time
+                       * @return the account information response from Binance
+                       * @throws sungjun.bitcoin.algorithmtrading.infrastructure.exception.binance.BinanceApiException if the API call fails
+                       */
     @GetExchange("/api/v3/account")
     BinanceAccountApiResponse getAccount(@RequestParam(value = "omitZeroBalances", required = false) boolean omitZeroBalances,
                       @RequestParam(value = "recvWindow", required = false) long recvWindow,
                       @RequestParam("timestamp") long timestamp);
 
     /**
-     * 계정 정보를 조회합니다 (잔고 필터링 및 타임스탬프 지정).
-     * <p>
-     * recvWindow는 기본값 5000ms로 설정됩니다.
-     * </p>
+     * Retrieves account information from Binance with an option to omit zero balances, using a default request validity window of 5000 milliseconds.
      *
-     * @param omitZeroBalances 잔고가 0인 자산을 응답에서 제외할지 여부
-     * @param timestamp 요청 시각 (밀리초 단위 Unix timestamp)
-     * @return {@link BinanceAccountApiResponse} 계정 정보 응답
+     * @param omitZeroBalances whether to exclude assets with zero balance from the response
+     * @param timestamp the Unix timestamp (in milliseconds) representing the request time
+     * @return the account information response from Binance
      */
     default BinanceAccountApiResponse getAccount(boolean omitZeroBalances, long timestamp) {
         return getAccount(omitZeroBalances, 5000L, timestamp); // recvWindow 기본값 5000L
     }
 
     /**
-     * 계정 정보를 조회합니다 (수신 윈도우 및 타임스탬프 지정).
-     * <p>
-     * omitZeroBalances는 false로 설정되어 모든 자산 정보를 포함합니다.
-     * </p>
+     * Retrieves account information from Binance with all asset balances included.
      *
-     * @param recvWindow 요청 유효 시간 (밀리초, 최대 60000)
-     * @param timestamp 요청 시각 (밀리초 단위 Unix timestamp)
-     * @return {@link BinanceAccountApiResponse} 계정 정보 응답
+     * The request uses the specified validity window and timestamp. All assets are included, even those with zero balances.
+     *
+     * @param recvWindow The request validity window in milliseconds (maximum 60000).
+     * @param timestamp The Unix timestamp in milliseconds representing the request time.
+     * @return The account information response from Binance.
      */
     default BinanceAccountApiResponse getAccount(long recvWindow, long timestamp) {
         return getAccount(false, recvWindow, timestamp); // omitZeroBalances false
     }
 
     /**
-     * 계정 정보를 조회합니다 (타임스탬프만 지정).
-     * <p>
-     * 기본 설정으로 모든 자산 정보를 포함하고 recvWindow는 5000ms로 설정됩니다.
-     * </p>
+     * Retrieves account information using the specified timestamp, including all assets and a default request window of 5000 ms.
      *
-     * @param timestamp 요청 시각 (밀리초 단위 Unix timestamp)
-     * @return {@link BinanceAccountApiResponse} 계정 정보 응답
+     * @param timestamp Unix timestamp in milliseconds representing the request time.
+     * @return BinanceAccountApiResponse containing account details.
      */
     default BinanceAccountApiResponse getAccount(long timestamp) {
         return getAccount(false, 5000L, timestamp); // 기본값 설정
