+++ src/main/java/sungjun/bitcoin/algorithmtrading/infrastructure/client/coinone/CoinoneTickerApiClient.java
@@ -21,15 +21,12 @@ import sungjun.bitcoin.algorithmtrading.infrastructure.client.coinone.response.C
 public interface CoinoneTickerApiClient {
 
     /**
-     * 지정된 통화쌍의 시세 정보를 조회합니다.
-     * <p>
-     * Coinone의 새로운 ticker API를 사용하여 실시간 시세 데이터를 가져옵니다.
-     * </p>
+     * Retrieves real-time ticker information for the specified currency pair from the Coinone exchange.
      *
-     * @param quoteCurrency 기준 통화 코드 (예: KRW)
-     * @param targetCurrency 대상 통화 코드 (예: BTC)
-     * @return {@link CoinoneTickerApiResponse} 시세 정보 응답
-     * @throws sungjun.bitcoin.algorithmtrading.infrastructure.client.coinone.exception.CoinoneApiException API 호출 실패 시
+     * @param quoteCurrency the base currency code (e.g., KRW)
+     * @param targetCurrency the target cryptocurrency code (e.g., BTC)
+     * @return the ticker data for the given currency pair
+     * @throws sungjun.bitcoin.algorithmtrading.infrastructure.exception.coinone.CoinoneApiException if the API call fails
      */
     @GetExchange("/ticker_new/{quoteCurrency}/{targetCurrency}")
     CoinoneTickerApiResponse getTicker(@PathVariable String quoteCurrency, @PathVariable String targetCurrency);
