--- src/main/java/sungjun/bitcoin/algorithmtrading/infrastructure/client/coinone/request/CoinoneOrderCancleRequest.java
@@ -1,19 +1,0 @@
-package sungjun.bitcoin.algorithmtrading.infrastructure.client.coinone.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class CoinoneOrderCancleRequest extends CoinonePrivateBaseRequest {
-
-    private String quoteCurrency;
-
-    private String targetCurrency;
-
-    @Builder
-    private CoinoneOrderCancleRequest(String accessToken, String nonce, String quoteCurrency, String targetCurrency) {
-        super(accessToken, nonce);
-        this.quoteCurrency = quoteCurrency;
-        this.targetCurrency = targetCurrency;
-    }
-}
