--- src/main/java/sungjun/bitcoin/algorithmtrading/infrastructure/client/interceptor/LoggingInterceptor.java
@@ -1,68 +1,0 @@
-package sungjun.bitcoin.algorithmtrading.infrastructure.client.interceptor;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpRequest;
-import org.springframework.http.client.ClientHttpRequestExecution;
-import org.springframework.http.client.ClientHttpRequestInterceptor;
-import org.springframework.http.client.ClientHttpResponse;
-import org.springframework.stereotype.Component;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-
-/**
- * HTTP 요청과 응답을 로그로 기록하는 인터셉터입니다.
- * <p>
- * 모든 외부 API 호출에 대해 요청 URL, 메서드, 헤더, 본문과
- * 응답 상태 코드, 헤더, 본문을 상세하게 로그로 기록합니다.
- * 디버깅 및 문제 해결에 유용합니다.
- * </p>
- *
- * @author sungjun
- * @since 1.0
- */
-@Slf4j
-@Component
-public class LoggingInterceptor implements ClientHttpRequestInterceptor {
-
-    @Override
-    public ClientHttpResponse intercept(HttpRequest request, byte[] body, ClientHttpRequestExecution execution) throws IOException {
-        logRequest(request, body);
-
-        // 응답을 래핑하여 복사된 본문으로 처리
-        ClientHttpResponse response = execution.execute(request, body);
-
-        logResponse(response);
-
-        return response;
-    }
-
-    private void logRequest(HttpRequest request, byte[] body) {
-        log.info("===========================request begin================================================");
-        log.info("URI         : {}", request.getURI());
-        log.info("Method      : {}", request.getMethod());
-        log.info("Headers     : {}", request.getHeaders());
-        log.info("Request body: {}", new String(body, StandardCharsets.UTF_8));
-        log.info("==========================request end================================================");
-    }
-
-    private void logResponse(ClientHttpResponse response) throws IOException {
-        StringBuilder inputStringBuilder = new StringBuilder();
-        try (BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(response.getBody(), StandardCharsets.UTF_8))) {
-            String line = bufferedReader.readLine();
-            while (line != null) {
-                inputStringBuilder.append(line);
-                inputStringBuilder.append('\n');
-                line = bufferedReader.readLine();
-            }
-        }
-        log.info("============================response begin==========================================");
-        log.info("Status code  : {}", response.getStatusCode());
-        log.info("Status text  : {}", response.getStatusText());
-        log.info("Headers      : {}", response.getHeaders());
-        log.info("Response body: {}", inputStringBuilder);
-        log.info("=======================response end=================================================");
-    }
-}
