+++ src/test/java/sungjun/bitcoin/algorithmtrading/api/service/ClientTestConfiguration.java
@@ -2,7 +2,7 @@ package sungjun.bitcoin.algorithmtrading.api.service;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import sungjun.bitcoin.algorithmtrading.client.MockCoinoneTickerApiClient;
+import sungjun.bitcoin.algorithmtrading.infrastructure.client.coinone.MockCoinoneTickerApiClient;
 import sungjun.bitcoin.algorithmtrading.infrastructure.client.coinone.CoinoneTickerApiClient;
 import sungjun.bitcoin.algorithmtrading.service.CoinoneService;
