--- src/test/java/sungjun/bitcoin/algorithmtrading/client/IntegrationClientTestSupport.java
@@ -1,7 +1,0 @@
-package sungjun.bitcoin.algorithmtrading.client;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public abstract class IntegrationClientTestSupport {
-}
