+++ src/test/java/sungjun/bitcoin/algorithmtrading/infrastructure/client/IntegrationClientTestSupport.java
@@ -1,0 +1,7 @@
+package sungjun.bitcoin.algorithmtrading.infrastructure.client;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public abstract class IntegrationClientTestSupport {
+}
